+++ Items/Ring.cpp
@@ -65,7 +65,7 @@ MeshPoint Ring::makePt(double angle) {
    printf("u: %f, v: %f\n", u, v);
    */
    MeshPoint tmpPt = MeshPoint(tmpx, tmpy, tmpz);
-   tmpPt.setTexPoint(u, v);
+   tmpPt.setTexPoint((float) u, (float) v);
    return tmpPt;
 }
